updated to svelte 5, trying to add badge counter back
