+++ package.json
@@ -10,8 +10,8 @@
 	"devDependencies": {
 		"@sveltejs/adapter-auto": "^3.0.0",
 		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
-		"svelte": "^4.2.7",
+		"@sveltejs/vite-plugin-svelte": "^4.0.1",
+		"svelte": "^5.2.3",
 		"vite": "^5.0.3"
 	},
 	"type": "module"
