+++ src/lib/atom/CounterBadge.svelte
@@ -1,0 +1,5 @@
+<script>
+    export let count = 0;
+</script>
+
+<span>{count}</span>
